Create README.md
The purpose and audience for this app:
This is a customizable nightlight where the user can touch the screen to create various objects that then bounce 
around the screen and either disappear (in Play mode) or continue (in Nightlight mode) to bounce around.  Users would 
anyone who wants a pleasant, lightly entertaining means of relaxing into bedtime, or anyone who is tired of the static,
typical nightlight.